refactor layout of popular persons

+++ app/src/main/java/com/example/omar/manasattask/data/retrofit/pojo/popular/PopularResult.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class PopularResult(
     val id: Int,
     val name: String,
-    val popularity: Float,
+    val popularity: Double,
     @SerializedName("profile_path")val profilePath:String
 )
